requirements.txt
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
arrow==0.17.0
asgiref==3.4.1
astroid==2.3.3
async-timeout==4.0.3
attrs==19.3.0
autobahn==20.2.1
Automat==0.8.0
autopep8==1.4.4
Babel==2.8.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
bleach==6.0.0
blog==1.0.1
bootstrap4==0.1.0
bumpversion==0.5.3
certifi==2019.9.11
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
constantly==15.1.0
cookiecutter==1.7.2
crispy-bootstrap4==2022.1
cryptography==41.0.3
cycler==0.11.0
daphne==2.4.1
dateparser==0.7.2
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==2.1.0
Django==3.2.6
django-active-link==0.1.6
django-activity-stream==0.9.0
django-admin-view-permission==1.9
django-adminactions==2.1.0
django-allauth==0.56.1
django-bootstrap5==23.3
django-braces==1.14.0
django-chartit==0.2.9
django-ckeditor==6.7.0
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.2
django-discover-runner==1.0
django-filter==2.2.0
django-heroku==0.3.1
django-infinite-scroll-pagination==1.2.1
django-js-asset==2.1.0
django-model-utils==4.0.0
django-notifications==0.1.dev0
django-notifications-hq==1.6.0
django-online-users==0.3
django-private-chat==0.3.0
django-redis==5.3.0
django-registration-redux==2.7
django-rest-framework==0.1.0
django-role-permissions==2.2.1
django-simple-history==3.4.0
django-taggit==4.0.0
django-templated-mail==1.1.1
django-widget-tweaks==1.5.0
djangorestframework==3.11.0
djoser==2.0.3
docopt==0.6.2
docutils==0.16
fake-factory==9999.9.9
Faker==2.0.4
filelock==3.0.12
fonttools==4.31.2
gunicorn==20.0.4
hiredis==1.0.1
humanize==4.8.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==4.8.3
incremental==17.5.0
isort==4.3.21
jet-bridge-base==0.6.6
jet-django==0.8.9
Jinja2==2.11.2
jinja2-time==0.2.0
jsonfield==3.1.0
kiwisolver==1.4.1
lazy-object-proxy==1.4.3
Markdown==3.3.7
MarkupSafe==1.1.1
martor==1.6.28
matplotlib==3.5.1
mccabe==0.6.1
msgpack==0.6.2
numpy==1.22.3
oauthlib==3.2.2
packaging==20.4
Pillow==6.2.1
pluggy==0.13.1
poyo==0.5.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.7.2
PyHamcrest==2.0.0
PyJWT==2.8.0
pylint==2.4.4
pymdown-extensions==10.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2019.3
PyYAML==6.0.1
redis==5.0.0
regex==2019.11.1
requests==2.24.0
requests-oauthlib==1.3.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.5
Sphinx==3.0.3
sphinx-rtd-theme==0.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.11
sqlparse==0.4.1
swapper==1.1.2.post1
text-unidecode==1.3
toml==0.10.2
tox==3.20.1
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
typing_extensions==4.7.1
tzdata==2022.1
tzlocal==2.0.0
urllib3==1.25.7
virtualenv==20.1.0
wcwidth==0.1.7
webencodings==0.5.1
websockets==8.1
whitenoise==6.5.0
wrapt==1.11.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
zope.interface==4.7.1
